Refactor quantization utility functions in test_quantization.py
### What's changed
- Moved utility functions for quantization, including `convert_scalar_to_ttnn_tensor`, `calculate_scale_zero_point_per_tensor`, and `check_match_ratio`, to a separate import statement for better organization.
- Added the `calculate_scale_zero_point_per_channel` function to enhance per-channel quantization capabilities.
- Removed redundant definitions of utility functions to streamline the code.

### Checklist
- [ ] Code is organized and functions are properly imported
- [ ] New functionality for per-channel quantization is implemented
